requirements.txt
Flask==2.1.2
python_dateutil==2.8.2
pytz==2022.1
requests==2.28.1
schedule==1.1.0
todoist==0.0.1
todoist_python==8.1.3
